@jonathanpeppers 2019-05-21
Build logs: [SmartHotel.zip](https://github.com/xamarin/xamarin-android/files/3203706/SmartHotel.zip)