@jonathanpeppers 2019-05-21
I think the build failures look OK. Windows has AOT failures, and macOS has network-related failures.